Able to get critics settings from backend
