+++ Arkumida/webapi/Models/Api/DTOs/Creatures/CreatureWithProfileDto.cs
@@ -19,7 +19,7 @@
 using System.Text.Json.Serialization;
 using webapi.Models.Creatures;
 
-namespace webapi.Models.Api.DTOs;
+namespace webapi.Models.Api.DTOs.Creatures;
 
 /// <summary>
 /// Creature with profile DTO
