+++ Arkumida/webapi/Models/Api/DTOs/TextDto.cs
@@ -17,6 +17,7 @@
 #endregion
 
 using System.Text.Json.Serialization;
+using webapi.Models.Api.DTOs.Creatures;
 
 namespace webapi.Models.Api.DTOs;
