+++ Arkumida/webapi/Models/Api/Responses/Creatures/InitiateEmailChangeResponse.cs
@@ -17,35 +17,33 @@
 #endregion
 
 using System.Text.Json.Serialization;
-using webapi.Models.Api.DTOs;
-using webapi.Models.Api.DTOs.Creatures;
 
 namespace webapi.Models.Api.Responses;
 
 /// <summary>
-/// Response for "find creature by name" request
+/// Result of email change initiation
 /// </summary>
-public class FindCreatureByNameResponse
+public class InitiateEmailChangeResponse
 {
     /// <summary>
-    /// Is creature found?
+    /// Is confirmation initiated successfully?
     /// </summary>
-    [JsonPropertyName("isFound")]
-    public bool IsFound { get; private set; }
+    [JsonPropertyName("isSuccessful")]
+    public bool IsSuccessful { get; private set; }
 
     /// <summary>
-    /// Creature (if found)
+    /// Is confirmation email sent (we wouldn't sent it in case of change to an empty email)?
     /// </summary>
-    [JsonPropertyName("creature")]
-    public CreatureDto Creature { get; private set; }
+    [JsonPropertyName("isEmailSent")]
+    public bool IsEmailSent { get; private set; }
 
-    public FindCreatureByNameResponse
+    public InitiateEmailChangeResponse
     (
-        bool isFound,
-        CreatureDto creature
+        bool isSuccessful,
+        bool isEmailSent
     )
     {
-        IsFound = isFound;
-        Creature = creature; // May be null if not found
+        IsSuccessful = isSuccessful;
+        IsEmailSent = isEmailSent;
     }
 }
