+++ Arkumida/webapi/Models/Api/Responses/CreaturesWithProfilesListResponse.cs
@@ -18,6 +18,7 @@
 
 using System.Text.Json.Serialization;
 using webapi.Models.Api.DTOs;
+using webapi.Models.Api.DTOs.Creatures;
 
 namespace webapi.Models.Api.Responses;
